requirements.txt
Flask==3.0.3
gunicorn==23.0.0
requests==2.32.3
google-generativeai==0.8.3
Pillow==10.4.0
flask-cors==4.0.1
python-dotenv==1.0.0